Icons for site navigation:
Add simple handlig to set color and icon for top level navigation items.
This is based on the title of the menu item.

This will need to be rewritten to use some kind of settings to set icon
and color. But it works as a quick solution for sundsvall.se because top
level navigation will probably not change.